Added file containing Global Land Temperatures by Country
There's a list of land temperatures by City that I found, but the size is half a gigabtyte of text, and It takes several minutes to load on the only text editor I could find that would even consider opening it. Instead, we will use this data to get the temp by country. There is not much data (that I could find given a $0 budget) on other weather statistics, so we'll make do with this.